refactor(empresa): extract use case from service, rename repos
Extract GetLastEmpresaUseCaseImpl from EmpresaService to properly separate use cases from the service layer in hexagonal architecture.

Changes:

- New GetLastEmpresaUseCaseImpl with UUID generation + save logic (moved from EmpresaService)
- EmpresaService now acts as an orchestrator delegating to GetLastEmpresaUseCase
- Rename EmpresaJpaRepository -> JpaEmpresaRepository and adapter for naming consistency
- EmpresaController injects EmpresaService instead of the use case directly
- Add comprehensive unit tests for GetLastEmpresaUseCaseImpl (3 cases: with businessId, null businessId generates UUID, empresa not found)
- Simplify EmpresaServiceTest to verify delegation only
- Bump version 2.3.0 -> 2.3.1 and update CHANGELOG, README, class diagram
