+++ pom.xml
@@ -11,7 +11,7 @@
     </parent>
     <groupId>com.termascacheuta</groupId>
     <artifactId>eterea.core.service</artifactId>
-    <version>2.3.0</version>
+    <version>2.3.1</version>
     <name>ETEREA.core-service</name>
     <description>API - Eterea - Core</description>
     <properties>
