+++ src/main/java/eterea/core/service/hexagonal/empresa/infrastructure/web/controller/EmpresaController.java
@@ -1,6 +1,6 @@
 package eterea.core.service.hexagonal.empresa.infrastructure.web.controller;
 
-import eterea.core.service.hexagonal.empresa.domain.ports.in.GetLastEmpresaUseCase;
+import eterea.core.service.hexagonal.empresa.application.service.EmpresaService;
 import eterea.core.service.hexagonal.empresa.infrastructure.web.dto.EmpresaResponse;
 import eterea.core.service.hexagonal.empresa.infrastructure.web.mapper.EmpresaDtoMapper;
 import lombok.RequiredArgsConstructor;
@@ -14,12 +14,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class EmpresaController {
 
-    private final GetLastEmpresaUseCase getLastEmpresaUseCase;
+    private final EmpresaService empresaService;
     private final EmpresaDtoMapper empresaDtoMapper;
 
     @GetMapping("/top")
     public ResponseEntity<EmpresaResponse> findTop() {
-        return getLastEmpresaUseCase
+        return empresaService
                 .findLast()
                 .map(empresa -> ResponseEntity.ok(empresaDtoMapper.toResponse(empresa)))
                 .orElse(ResponseEntity.notFound().build());
