+++ src/test/java/eterea/core/service/hexagonal/empresa/application/service/EmpresaServiceTest.java
@@ -1,7 +1,7 @@
 package eterea.core.service.hexagonal.empresa.application.service;
 
 import eterea.core.service.hexagonal.empresa.domain.model.Empresa;
-import eterea.core.service.hexagonal.empresa.domain.ports.out.EmpresaRepository;
+import eterea.core.service.hexagonal.empresa.domain.ports.in.GetLastEmpresaUseCase;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -9,55 +9,29 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Optional;
-import java.util.UUID;
 
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class EmpresaServiceTest {
 
     @Mock
-    private EmpresaRepository empresaRepository;
+    private GetLastEmpresaUseCase getLastEmpresaUseCase;
 
     @InjectMocks
     private EmpresaService empresaService;
 
     @Test
-    void findLast_WhenEmpresaExistsAndBusinessIdIsNotNull_ShouldReturnEmpresa() {
-        UUID businessId = UUID.randomUUID();
-        Empresa empresa = Empresa.builder().businessId(businessId).build();
-
-        when(empresaRepository.findLast()).thenReturn(Optional.of(empresa));
+    void findLast_ShouldDelegateToUseCase() {
+        Empresa empresa = Empresa.builder().build();
+        when(getLastEmpresaUseCase.findLast()).thenReturn(Optional.of(empresa));
 
         Optional<Empresa> result = empresaService.findLast();
 
-        assertTrue(result.isPresent());
-        assertEquals(businessId, result.get().getBusinessId());
-        verify(empresaRepository, never()).save(any());
+        assertEquals(Optional.of(empresa), result);
+        verify(getLastEmpresaUseCase).findLast();
     }
 
-    @Test
-    void findLast_WhenEmpresaExistsAndBusinessIdIsNull_ShouldGenerateUuidAndSave() {
-        Empresa empresa = Empresa.builder().businessId(null).build();
-
-        when(empresaRepository.findLast()).thenReturn(Optional.of(empresa));
-
-        Optional<Empresa> result = empresaService.findLast();
-
-        assertTrue(result.isPresent());
-        assertNotNull(result.get().getBusinessId());
-        verify(empresaRepository).save(empresa);
-    }
-
-    @Test
-    void findLast_WhenEmpresaDoesNotExist_ShouldReturnEmpty() {
-        when(empresaRepository.findLast()).thenReturn(Optional.empty());
-
-        Optional<Empresa> result = empresaService.findLast();
-
-        assertTrue(result.isEmpty());
-        verify(empresaRepository, never()).save(any());
-    }
 }
